media/categories add no interplay

+++ media_site/src/main/java/lt/techin/media_site/repository/RoleRepository.java
@@ -3,6 +3,6 @@ package lt.techin.media_site.repository;
 import lt.techin.media_site.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role,Long> {
+public interface RoleRepository extends JpaRepository<Role,Byte> {
     Role findByName(String name);
 }
